Added HTTPS support
The "hc_" hack was generalized to also work for https connections.  In
addition, there now exists a TLSClientConfig object within the
HttpClient object that can be used to control the TLS settings for HTTPS
connections.  One caviat is that the object cannot be moved anywhere in
memory, so it can only be changed inplace.